moved inventory data loading logic to inventory/+layout.svelte
